add modal for deleting months

+++ front-end-app/src/app/month/month.component.ts
@@ -1,6 +1,7 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {MonthDataService} from "../services/month-data.service";
 import {FlashMessagesService} from "angular2-flash-messages";
+import {ConfirmationDialogService} from "../confirmation-dialog/confirmation-dialog.service";
 
 @Component({
   selector: 'app-month',
@@ -13,28 +14,43 @@ export class MonthComponent implements OnInit {
   @Input() urlBase: string;
   @Output() isChanged = new EventEmitter<boolean>();
 
-  constructor(private monthDataService: MonthDataService, private flashMessages: FlashMessagesService,) {
+  constructor(private monthDataService: MonthDataService, private flashMessages: FlashMessagesService, private confirmationDialogService: ConfirmationDialogService) {
   }
 
   ngOnInit(): void {
   }
 
-  deleteMonth(event, month): void {
-    this.monthDataService.deleteMonth(this.urlBase+'/'+month).subscribe(({msg, success}) => {
-      if (!success) {
-        this.flashMessages.show(msg, {
-          cssClass: 'alert-danger',
-          timeout: 2000
-        })
+  async deleteMonth(event, month): Promise<void> {
+    const modalParams = [
+      'Подтввердите удаление..',
+      `Удалить анные за период ${month}?`,
+      'УДАЛИТЬ',
+      'ОТМЕНИТЬ',
+      'lg'
+    ]
 
-      } else {
-        this.flashMessages.show(msg, {
-          cssClass: 'custom-success',
-          timeout: 2000
-        })
-        this.isChanged.emit(true);
-      }
-    });
+    // @ts-ignore
+    const confirmed = await this.confirmationDialogService.confirm(...modalParams);
+    if (confirmed) {
+      this.monthDataService.deleteMonth(this.urlBase + '/' + month).subscribe(({msg, success}) => {
+        if (!success) {
+          this.flashMessages.show(msg, {
+            cssClass: 'alert-danger',
+            timeout: 2000
+          })
+
+        } else {
+          this.flashMessages.show(msg, {
+            cssClass: 'custom-success',
+            timeout: 2000
+          })
+          this.isChanged.emit(true);
+        }
+      });
+
+    }
+
+
     return;
   }
 }
